Adds link recognition to tabs
The tab directive uses anchor tags to represent the tabs. Since there is no `href` attribute, browsers do not recognize them as so. This prevents users from being able to interact with the tabs fully as if they were. This includes features such as being able to middle click a link to open in a new tab.
This change adds a blank `href` attribute as a hook for browsers. Since it is blank, it will not interfere with usual routing.`