feat(drilldownMenu): Add automatic resizing, an API, and events.
This updates adds 3 key updates to the existing drilldownMenu:
- automatic resizing when the viewport is resized.
- an API to programmatically control some features.
- a set of events, emitted as appropriate.

Automatic resizing was needed because the component only set its sizing
when it was initialised.  If the size of the screen became narrower (e.g.
landscape to portrait on a phone), the menus could easily be too large
for the initial area.

The API provides methods to resize and open and close submenus.  The
resize() method is particularly important in cases where the menu is
dynamically shown or hidden with ng-show.  See the updated readme for
details.

There are also a few events that are emitted to allow parent controllers
etc. to be notified. See the update readme for details.

Finally, this revision also fixes some memory leaks that happened when
the component was destroyed (e.g. via ng-if).

Test Plan:
- Ensure the unit tests still run and pass

- Test the API functions through the newly added demo section and
  ensure they affect the component as appropriate.
- Test the events are correctly emitted, both when the condition is
  triggered by normal usage, and when triggered by the API
- Test the control doesn't leak when used with ng-if (replacing the ng-show
  in the API demo section).
- Check the above works on:
  - Windows Chrome, Firefox, Edge, and IE 11
  - iOS Safari