fix(drilldownMenu): Move API initialisation to $onInit()
The API was being initialised in the main controller function, but
it should have been set in `$onInit()`.  In some cases the api was
being overwritten by the value from the parent scope, rather than
the other way round.

Not sure exactly why this didn't happen in the demo code, but did
happen in my actual app.  In any case, $onInit() is the correct place
as it is run after bindings have been initialised.

Test Plan:
- Check this still works in the demo page.
- Check this works in my own application.