feat: Support content security policy in the local serving of the demo page
## Summary

- Adds `--csp` flag to gulp to send CSP headers via the browsersync server
- Adds [`ng-csp`](https://code.angularjs.org/1.6.5/docs/api/ng/directive/ngCsp) and related css
  for Angular's internal csp support.
- Updates the demo page generation to write the demo js to file rather
  than inline in the demo page to avoid inline script issues

WARNING: most directives on the demo page DO NOT currently work with CSP enabled!

## Details

[Content security policy](https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP)
is a useful security layer to protect applications against XSS and
other attacks.

It usually prevents use of inline scripts and inline styles, which
up to now have been used in a number of places. So its important to
test with CSP enabled to ensure the directives are usable
in CSP secured pages.

This revision makes it easy to test by providing a `--csp` flag that
will send a minimal set of content-security-policy headers to the
browser as part of serving the demo page. Ideally CSP would be enabled
by default, but too many of the directives and demos would break at
this time.

The configured policy does not include `unsafe-inline` (for obvious
reasons) so this change also had to remove the inlining of scripts into
the demo page.  This is replaced by generating an `index.js` file
and sourcing that in the index.html instead (which is secure as it comes
from `'self'`).

The csp support comes from [helmet-csp](https://helmetjs.github.io/docs/csp/),
a well regarded security middleware.

## Test Plan
- Run `gulp`
-- Check all unit test still run and pass
-- Check the demo page is generated and shows in the browser
-- Check the response headers DO NOT include content-security-policy
-- Check all directives continue to work as before
-- Check no errors are shown in the console

- Run `gulp --csp`
-- Check all unit test still run and pass
-- Check the demo page is generated and shows in the browser
-- Check the response headers DO include content-security-policy
-- Check at least some directives continue to work
-- Check all fonts and external scripts are allowed to be loaded
-- Check the console shows CSP-related errors for the directives or
   demos that don't work.