fix(demo): fix the demo on gh-pages by including the index.js
As described in #72, the fix in #66 added a new file for the demo pages: `dist/index.js`.
This wasn't being carried through to the gh-pages because the `publish` task commits
updated files, but not added ones.

So this diff just adds the latest version of index.js so that it will be picked up by the
`publish` task in future.

## Test Plan
Serve the `dist ` folder, and check it works correctly in the browser.